Move glwegadget functions
