+++ benchmarks/bench_ggsw_external_prod.cpp
@@ -7,7 +7,6 @@ extern "C" {
 #include "ggsw_arithmetic.h"
 #include "ggsw_ciphertext.h"
 #include "ggsw_params.h"
-#include "glwe_arithmetic.h"
 #include "glwe_params.h"
 #include "glwe_transform_key.h"
 #include "rng.h"
