+++ benchmarks/bench_glwegad_half_prod.cpp
@@ -6,9 +6,8 @@ extern "C" {
 #include "bivariate_polynomial.h"
 #include "ggsw_arithmetic.h"
 #include "ggsw_params.h"
-#include "glwe_arithmetic.h"
 #include "glwe_transform_key.h"
-#include "glwegadget.h"
+#include "glwegadget_arithmetic.h"
 #include "glwegadget_ciphertext.h"
 #include "rng.h"
 #include "univariate_polynomial.h"
