+++ include/core/ggsw/glwegadget_ciphertext.h
@@ -3,6 +3,7 @@
 
 #include "bivariate_polynomial.h"
 #include "ggsw_params.h"
+#include "glwe_key.h"
 
 /**
  * @brief A GLWEGadget ciphertext
@@ -67,4 +68,56 @@ GLWEGadgetCiphertextPrep* new_glwegadget_prep(const GLWEGadgetParams* params);
  */
 void delete_glwegadget_prep(GLWEGadgetCiphertextPrep* glwegadget_prep_ct);
 
+/**
+ * @brief Encrypts a univariate integer polynomial into a GLWEGadget
+ * (also known as halfGGSW in the k=1 case)
+ *
+ * @param module The underlying compute module
+ * @param result The resulting GLWEGadget ciphertext
+ * @param sk_dft A prepared secret key
+ * @param m_univ The univariate plaintext
+ *
+ * @retval -1 if an error occurs
+ * @retval 0 otherwise
+ */
+int glwegadget_secret_encrypt(const MODULE* module, GLWEGadgetCiphertext* result, const GLWESecretKeyDFT* sk_dft,
+                              const PolyUniv* m_univ);
+
+/**
+ * @brief Encrypts a univariate integer polynomial into a GLWEGadget
+ * (also known as halfGGSW in the k=1 case)
+ *
+ *  TODO: implement
+ *
+ * @param module The underlying compute module
+ * @param result The resulting GLWEGadget ciphertext
+ * @param pk     A public key
+ * @param m_univ The univariate plaintext
+ *
+ */
+void glwegadget_public_encrypt(const MODULE* module, GLWEGadgetCiphertext* result, const GLWEPublicKey* pk,
+                               const PolyUniv* m_univ);
+
+/**
+ * @brief Retreive a bivglwe from a GLWEGadget (halfGGSW for k = 1)
+ *
+ * @param glwegadget_ct     The GLWEGadget ciphertext
+ * @param i                 The index of the GLWE to retrievie, from 1 to l_tilde included
+ *
+ */
+VecBiv* glwegadget_extract_bivglwe(GLWEGadgetCiphertext* glwegadget_ct, uint64_t i);
+
+/**
+ * @brief Prepares a GLWEGadget for use in a half-external product
+ *
+ * @param module The underlying backend module
+ * @param glwegadget_prep_ct The resulting prepared GLWEGadget
+ * @param glwegad_ct The input GLWEGadget
+ *
+ * @retval -1 if an error occurs
+ * @retval 0 otherwise
+ *
+ */
+int glwegadget_prepare(const MODULE* module, GLWEGadgetCiphertextPrep* glwegadget_prep_ct,
+                       const GLWEGadgetCiphertext* glwegad_ct);
 #endif  // PARTIALGGSW_CIPHERTEXT_H
