+++ tests/unit/core/ggsw/test_glwegadget_encrypt.c
@@ -4,7 +4,7 @@
 
 #include "bivariate_polynomial.h"
 #include "core/ggsw/ggsw_arithmetic.h"
-#include "core/ggsw/glwegadget.h"
+#include "core/ggsw/glwegadget_arithmetic.h"
 #include "core/glwe/glwe_arithmetic.h"
 #include "ggsw_params.h"
 #include "glwe_ciphertext.h"
